This commit contains several modifications to allow the Canvas Data Loader project to be able to better handle numerous data dump retrievals targeted by date, and also allow the dump metadata and download path locations to be made available to data processing and ingestion handlers for the Matthews LRW.
Refactored the Canvas Data API Client to use RestTemplate instead of JAX-RS
(Jersey). This was done for both personal preference and consistency, since
the MatthewsClient is already using RestTemplate.

Renamed the ambiguous ApiClient class to the more descriptive
CanvasDataApiClient to clearly separate it from the MatthewsClient. Also
merged all of the Canvas API handling into the CanvasDataApiClient and
eliminated the RestUtils which was not generically useful as a REST
client, it was very Canvas Data API specific.

Added a RestTemplate bean to the context explicitly with the standard
defaults generated by the RestTemplateBuilder. I had used this with some
experimental configuration customizations that I have since removed, but I think
keeping it defined as a bean with the Builder is useful in case there is future
need for customization.

Refactored Date handling to use the Java 8 Date Time APIs to allow for easier
handling of UTC dates. Canvas shares all dates and times in UTC.

Added an explicit Bean declaration for the Jackson ObjectMapper in order to
configure the Date/Time serialization to output ISO 8601 by default for all JSON
timestamps.

Created and added an EpochMillisecondsDeserializer type to properly convert
the Canvas Dump expired field which is sent as milliseconds since the epoch.

Added output of a dump-metadata.json file containing the Canvas dump metadata to
each dump directory. This is essentially the response payload (as it is mapped
based on what currently exists) from each detailed Canvas Data API dump call.
This is generally useful to provide context to the dump artifacts if the
processing is not done immediately after the download where the
CanvasDataDumps can be passed, instead we can add support for them to be
reloaded from disk. This allows for some dynamic data loading support for the
Matthews processors without having to infer based solely on the file system
contents, particularly since not all dump artifacts are consistent between dumps
(e.g. requests are not included in every dump).

Changed the environment property used to designate the artifact download
directory. This now designates the parent directory for dumps which will be
sub-directories dated YYYY-MM-DD (UTC time zone).

   OLD: downloaddirectory
   NEW: downloads.root.directory