Initial commit of support for reading the Canvas Quiz Submission Fact dump artifacts (both Current and Historical).
This solution leverages the robust (de-)serialization capabilities of Jackson
and its supporting libraries to read the tab delimited data dump files. It does
so by adding the jackson-dataformat-csv extension which supports changing the
delimiter from a comma to a tab.

To further simplify the implementation, Project Lombok has also been added to
handle generation of the boilerplate accessors, constructors and mutators, so
the new model type CanvasQuizSubmissionFact provides a very simple data
transfer object to represent all of the Quiz Submission Fact data.

As a result of adding Lombok, developers will need to enable or install
a plugin in IDEs to work effectively with the simplified model classes which
rely on auto-generated code. Maven builds without any changes required.
https://projectlombok.org/download.html

Simple and reusable JsonDeserializer extensions have been added to handle
common data deserialization requirements for the Canvas dump tabular data.