Added support for filtering during the deserialization process via a filter method which accepts an arbitrary Function (lambda expression). The lambda function must evaluate to true for an artifact item to be included.
Also added an includeOnly method on the reader to be able to constrain the
artifacts loaded which is useful for when a ReadableCanvasDumpArtifact
is actually able to represent multiple dump artifacts.

Example of filtering to only include items which have end dates, and
which are after a specified date.

    CanvasDataDumpReader.forType(CanvasQuizSubmissionFact.class)
        .withFilter(canvasQuizSubmissionFact ->
            canvasQuizSubmissionFact.getDate().isPresent() ?
            canvasQuizSubmissionFact.getDate().get().isAfter(
            LocalDate.parse("2016-10-21").atStartOfDay(
            ZoneOffset.UTC).toInstant()) : false)
        .read(dump);

Example of constraining the artifact types when a ReadableCanvasDumpArtifact
supports multiple:

    CanvasDataDumpReader.forType(CanvasQuizSubmissionFact.class)
        .includeOnly(CanvasQuizSubmissionFact.Types.quiz_submission_fact))
        .read(dump);

These methods may be chained.