This is initial work on the Canvas Page Requests conversion to Caliper.
This converter approach allows us to fairly cleanly load converters into
groups by the Source and Target types, and further filter within each of those
groups by the supports method in each converter which can examine the Source
model.

The Maps class is a shim based loosely on Java 9 Map.ofEntries which is useful
in conjunction with the fluent builder code.
Example:
    Map<String, String> map = Maps.ofEntries(
       entry("key1", "value1"),
       entry("key2", "value2"),
       entry("key3", "value3"));

Maps allows for type safe Map usage within fluent interfaces in a more seamless
way. I derived it from the blog post
http://minborgsjavapot.blogspot.com/2014/12/java-8-initializing-maps-in-smartest-way.html
and enhanced what that blog suggested to make it even simpler to use inline.
This is an interim utility until we get the new convenience factory methods of the same name
coming in Java 9 Collections.
http://download.java.net/java/jdk9/docs/api/java/util/Map.html#ofEntries-java.util.Map.Entry...-