Refinement to the Canvas Data Loader request conversion process.
After processing only a couple event types, it was obvious that some
refactoring to eliminate redundancy would be beneficial as we
significantly expand on the number of request type events being processed.

As a result, I added EventBuilderUtils to provide a wrapper around the
Event builder, allowing us to isolate the common (and in many cases
still uncertain) attributes so they are not replicated all over the
place. This distills each event converter down to the specific
attributes that are variable to each.

Due to the need for user and enrollment data in each event, I created
SupportingEntities as a simple holder for all of the supporting data.

This commit also adds parsing of the Canvas Pseudonym dimension data
via CanvasDataPseudonymDimension. We need that to get the user login
(pseudonym).

As mentioned in the code comments, transforming login events to Caliper
is not as straightforward as expected, since each actual Canvas event
record representing a login does not include the authenticating user
data. This commit adds an inefficient but necessary approach of scanning
the requests for another request in the same session which does contain
the user ID, and pulling that to send with the login event to Matthews.

NOTE: This is a work in progress. Refinement is still needed to insure
we are getting all of the login and logout events loaded properly for
the provided dumps. It is known that some are being omitted and I am
working on resolving those issues.