LAQ-163 Implement the Transform of Canvas Discussion Forum Entry created to IMS Caliper.
This includes adding the CanvasUserDimensions into the SupportingEntities
in order to get the root account ID for the user. This should probably
be rolled into a more comprehensive user model for the purpose of the
transform to reduce collection scanning, but we can do that later.

This also includes the old login scanning and needs to be improved to
leverage the userEmailMap.